PROJECT_WARNINGS

+++ src/main.cpp
@@ -1,10 +1,11 @@
+#include <cstddef>
 #include <da4cpp/parsing.hpp>
 #include <iostream>
 #include <optional>
 #include <span>
 
 int main(int argc, char **argv) {
-  const std::span<char *> args(argv, argc);
+  const std::span<char *> args(argv, static_cast<size_t>(argc));
   if (argc != 2) {
     std::cerr << "Usage: " << args[0] << " <source-file>" << std::endl;
     return 1;
